Full support for .gitignore files and patterns
This commit adds support for parsing and honoring exclusion
patterns when scanning directories.

It also adds a two new flags:  --ignore and --no-gitignore

When generating a package from a directory, bom will read the .gitignore
file at the top and apply the exclusions listed to the filesystem tree
unless --no-gitignore is set to true.

If more patterns are defined via --ignore, they will be appended
to the list found in .gitignore.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>