Correctly implement LicenseInfoFromFiles in spdx packages
This commit modifies the way LicenseInfoFromFiles is implemented
in SPDX. This field should be a list as can appear multiple
times in the package declartion.

In addition, we now enforce LicenseInfoFromFiles to be rendered
when FilesAnalyzed is true as demanded by the SPDX spec. If files
have no icensing info, LicenseInfoFromFiles is rendered correctly
with "NONE".

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>