Switch to license.ReadTopLicense()
This commit adds a new function to the license package: ReadTopLicense()

This function tries to determine the most significant license file
in a directory by first trying some common names at the top of the
tree and then working its way down looking for the topmost one.

The use of this function saves thousands of license scans and reduces
memory footprint when genearting the kubernetes sbom.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>