SPDX Relationships
This PR reworks the way we relate objects in the spdx package.
Instead of having a simple subpackage method, this commit
introduces a new Relationship type and a way to relate objects
among themselves via a new spdx.Object interface.

Both spdx.File and spdx.Package both embed a new type Entity which
implements most of the SPDX Object interface. This simplifies the
code a lot as I was able to remove a lot of redundant functions.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>