Add ExternalDocRef to SPDX package
This commit adds the capability to the spdx package to reference
external documents. This is achieved by a new type ExternalDocumentRef
which gets rendered in the document.

This commit also modifies the YAML configuration support to enable
the definition of external document references:

```yaml
external-docs:
  - id: kubernetes-source
    uri: http://gcr.io/k8s.spdx
    checksums: {sha256: 5167fec9c11112ee8ea3f2b21345df18d4a4aec5a5dcbeb03d0d2c8c438eada7}
```

A test for the correct rendering of the new type is included.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>