go pkg: provenance
This commit adds a new go package: provenance

This code contains the types and functions to generate provenance
metadata. It currently includes code to generate SLSA compliant
predicates and automatically scan direcories to add files as
subjects in a provenance attestation.

Signing is currently not supported but the envelope type is already
in the package for the next itearation, the push towards SLSA2 in
the Kubernetes release process.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>