bom: document outline subcommand
This commit adds a new subcommand to the bom utility: document outline.

When passing a SPDX document, bom will now render a tree visualization of
a bill of materials to help the user understand what is contained in it.

The initial document outline allows the user to specify the depth of
recursion and choose to use package names or SPDX identifiers when
rendering the outline.

The document subcommand will eventually harbor a `document verify` subcommand
which is still under construction.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>