Make documents queryable with GetElementByID()
This commit adds a new DOM-inspired function document.GetElementByID()
that allows a user to search in a document for a Package of file with a
given SPDX-ID.

If found, returns an instance of the object interface which can be
cast to check its type. Returns nil if no element matches the query.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>