GoMod: Use go proxy for dnload, fix locators
This commit modifies the DownloadLocation generated from Go packages.

When the package is a module, we will use the proxy.golang.org url
to make a zip available. This makes the output of bom equal to ko.

When a dependency is not a module, bom will now generate a valid SPDX
download locator (eg "git+https://github.com/ ... )

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>