Embed SPDX license list!
This commit adds code and data to embed the SPDX license list into the
bom binary. It speeds up SBOM generation considerably.

It also changes the handling of the catalog options to not use a pointer
to avoid overwriting the default options struct and suppot multiple
licenseList versions.

During normal operations, bom will use the embedded list, when requesting
another version, it will fetch it from github and use that.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>