Embed SPDX license list! (#255)
* Refactor license FS parser

This commit refactors the license reader to be reusable

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>

* Wrap errors in filesystem reader

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>

* License Downloader: Expose latest version method

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>

* magefile: Add license data checker

This commit adds a CHeckEmbeddedData command to the
magefile that checks that the license data is up to date.
It also consts the license data location in the license
package.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>

* Refactir license download to own  method

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>

* Downloader: Add Method to download license list to file

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>

* Magefile: UpdateEmbeddedData method

This commit adds a new method to the magefile to update the embedded data
in the license package. Using this method mage will download the SPDX zip
data and unpack the files we care about in the license directory to avoid
embedding the whole 52 MB of license data.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>

* Embed zip file instead of directory

This commit modifies the embed cose to embed a zip file
instead if the single files. This lowers the size by 80%

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>

* Embed SPDX license list!

This commit adds code and data to embed the SPDX license list into the
bom binary. It speeds up SBOM generation considerably.

It also changes the handling of the catalog options to not use a pointer
to avoid overwriting the default options struct and suppot multiple
licenseList versions.

During normal operations, bom will use the embedded list, when requesting
another version, it will fetch it from github and use that.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>

* Regenerate fakes

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>

* Update CheckEmbeddedData to work with zip and not dir

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>

---------

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>