Parser: Keep reading SBOM when detecting encoding
This commit modifies the encoding detection function to read
the as much of the SBOM as required to deetect the encoding.

Some tools (notably Microsoft's sbom-tool) put the document
creation data at the end of the document and we could not
detect the encoding by reading a few bytes from the top.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>